Reorganised API handlers
no refs

Reorganised API handlers to better represent the structure of the API:

- Used singular names for files containing API handlers
- Grouped together "Profile" related handlers in the same file
