+++ src/http/api/index.ts
@@ -1,9 +1,7 @@
 export * from './activities';
-export * from './accounts';
-export * from './feeds';
-export * from './follows';
-export * from './notes';
-export * from './posts';
-export * from './profiles';
+export * from './account';
+export * from './feed';
+export * from './note';
+export * from './profile';
 export * from './search';
-export * from './webhooks';
+export * from './webhook';
