+++ src/http/api/profile.ts
@@ -3,11 +3,235 @@ import { type Actor, type CollectionPage, isActor } from '@fedify/fedify';
 import type { AccountService } from '../../account/account.service';
 import { type AppContext, fedify } from '../../app';
 import {
+    getAttachments,
+    getFollowerCount,
+    getFollowingCount,
+    getHandle,
     isFollowedByDefaultSiteAccount,
     isHandle,
 } from '../../helpers/activitypub/actor';
+import { sanitizeHtml } from '../../helpers/html';
 import { isUri } from '../../helpers/uri';
 import { lookupObject } from '../../lookup-helpers';
+
+interface Profile {
+    actor: any;
+    handle: string;
+    followerCount: number;
+    followingCount: number;
+    isFollowing: boolean;
+}
+
+/**
+ * Create a handler for a request for a profile
+ *
+ * @param accountService Account service instance
+ */
+export function createGetProfileHandler(accountService: AccountService) {
+    /**
+     * Handle a request for a profile
+     *
+     * @param ctx App context instance
+     */
+    return async function handleGetProfile(ctx: AppContext) {
+        const db = ctx.get('db');
+        const logger = ctx.get('logger');
+        const site = ctx.get('site');
+        const apCtx = fedify.createContext(ctx.req.raw as Request, {
+            db,
+            globaldb: ctx.get('globaldb'),
+            logger,
+        });
+
+        // Parse "handle" from request parameters
+        // /profile/:handle
+        const handle = ctx.req.param('handle') || '';
+
+        // If the provided handle is invalid, return early
+        if (isHandle(handle) === false) {
+            return new Response(null, { status: 400 });
+        }
+
+        // Lookup actor by handle
+        const result: Profile = {
+            actor: {},
+            handle: '',
+            followerCount: 0,
+            followingCount: 0,
+            isFollowing: false,
+        };
+
+        try {
+            const actor = await lookupObject(apCtx, handle);
+
+            if (!isActor(actor)) {
+                return new Response(null, { status: 404 });
+            }
+
+            result.actor = await actor.toJsonLd();
+            result.actor.summary = sanitizeHtml(result.actor.summary);
+            result.handle = getHandle(actor);
+
+            const [
+                followerCount,
+                followingCount,
+                isFollowingResult,
+                attachments,
+            ] = await Promise.all([
+                getFollowerCount(actor),
+                getFollowingCount(actor),
+                isFollowedByDefaultSiteAccount(actor, site, accountService),
+                getAttachments(actor, {
+                    sanitizeValue: (value: string) => sanitizeHtml(value),
+                }),
+            ]);
+
+            result.followerCount = followerCount;
+            result.followingCount = followingCount;
+            result.isFollowing = isFollowingResult;
+            result.actor.attachment = attachments;
+        } catch (err) {
+            logger.error('Profile retrieval failed ({handle}): {error}', {
+                handle,
+                error: err,
+            });
+
+            return new Response(null, { status: 500 });
+        }
+
+        // Return results
+        return new Response(JSON.stringify(result), {
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            status: 200,
+        });
+    };
+}
+
+interface PostsResult {
+    posts: any[];
+    next: string | null;
+}
+
+/**
+ * Handle a request for a profile's posts
+ *
+ * @param ctx App context instance
+ */
+export async function handleGetProfilePosts(ctx: AppContext) {
+    const db = ctx.get('db');
+    const logger = ctx.get('logger');
+    const apCtx = fedify.createContext(ctx.req.raw as Request, {
+        db,
+        globaldb: ctx.get('globaldb'),
+        logger,
+    });
+
+    // Parse "handle" from request parameters
+    // /profile/:handle/posts
+    const handle = ctx.req.param('handle') || '';
+
+    // If the provided handle is invalid, return early
+    if (!isHandle(handle)) {
+        return new Response(null, { status: 400 });
+    }
+
+    // Parse "next" from query parameters
+    // /profile/:handle/posts?next=<string>
+    const queryNext = ctx.req.query('next') || '';
+    const next = queryNext ? decodeURIComponent(queryNext) : '';
+
+    // If the next parameter is not a valid URI, return early
+    if (next !== '' && !isUri(next)) {
+        return new Response(null, { status: 400 });
+    }
+
+    // Lookup actor by handle
+    const actor = await lookupObject(apCtx, handle);
+
+    if (!isActor(actor)) {
+        return new Response(null, { status: 404 });
+    }
+
+    // Retrieve actor's posts
+    // If a next parameter was provided, use it to retrieve a specific page of
+    // posts. Otherwise, retrieve the first page of posts
+    const result: PostsResult = {
+        posts: [],
+        next: null,
+    };
+
+    let page: CollectionPage | null = null;
+
+    try {
+        if (next !== '') {
+            // Ensure the next parameter is for the same host as the actor. We
+            // do this to prevent blindly passing URIs to lookupObject (i.e next
+            // param has been tampered with)
+            // @TODO: Does this provide enough security? Can the host of the
+            // actor be different to the host of the actor's followers collection?
+            const { host: actorHost } = actor?.id || new URL('');
+            const { host: nextHost } = new URL(next);
+
+            if (actorHost !== nextHost) {
+                return new Response(null, { status: 400 });
+            }
+
+            page = (await lookupObject(apCtx, next)) as CollectionPage | null;
+
+            // Explicitly check that we have a valid page seeming though we
+            // can't be type safe due to lookupObject returning a generic object
+            if (!page?.itemIds) {
+                page = null;
+            }
+        } else {
+            const outbox = await actor.getOutbox();
+
+            if (outbox) {
+                page = await outbox.getFirst();
+            }
+        }
+    } catch (err) {
+        logger.error('Error getting outbox', { error: err });
+    }
+
+    if (!page) {
+        return new Response(null, { status: 404 });
+    }
+
+    // Return result
+    try {
+        for await (const item of page.getItems()) {
+            const activity = (await item.toJsonLd({
+                format: 'compact',
+            })) as any;
+
+            if (activity?.object?.content) {
+                activity.object.content = sanitizeHtml(activity.object.content);
+            }
+
+            if (typeof activity.actor === 'string') {
+                activity.actor = await actor.toJsonLd({ format: 'compact' });
+            }
+
+            result.posts.push(activity);
+        }
+    } catch (err) {
+        logger.error('Error getting posts', { error: err });
+    }
+
+    result.next = page.nextId
+        ? encodeURIComponent(page.nextId.toString())
+        : null;
+
+    return new Response(JSON.stringify(result), {
+        headers: {
+            'Content-Type': 'application/json',
+        },
+        status: 200,
+    });
+}
 
 interface FollowersResult {
     followers: {
@@ -22,13 +246,15 @@ interface FollowersResult {
  *
  * @param accountService Account service instance
  */
-export function createGetFollowersHandler(accountService: AccountService) {
+export function createGetProfileFollowersHandler(
+    accountService: AccountService,
+) {
     /**
      * Handle a request for a profile's followers
      *
      * @param ctx App context instance
      */
-    return async function handleGetFollowers(ctx: AppContext) {
+    return async function handleGetProfileFollowers(ctx: AppContext) {
         const db = ctx.get('db');
         const logger = ctx.get('logger');
         const site = ctx.get('site');
@@ -141,6 +367,7 @@ export function createGetFollowersHandler(accountService: AccountService) {
         });
     };
 }
+
 interface FollowingResult {
     following: {
         actor: any;
@@ -154,13 +381,15 @@ interface FollowingResult {
  *
  * @param accountService Account service instance
  */
-export function createGetFollowingHandler(accountService: AccountService) {
+export function createGetProfileFollowingHandler(
+    accountService: AccountService,
+) {
     /**
      * Handle a request for a profile's following
      *
      * @param ctx App context instance
      */
-    return async function handleGetFollowing(ctx: AppContext) {
+    return async function handleGetProfileFollowing(ctx: AppContext) {
         const db = ctx.get('db');
         const logger = ctx.get('logger');
         const site = ctx.get('site');
